pyproject.toml
[tool.poetry]
name = "ai-translate"
version = "0.1.0"
description = ""
authors = ["Elias Soltani <elias.soltani@fuel50.com>"]
license = "Properietary"
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
langchain = "^0.1.16"
openpyxl = "^3.1.2"
openai = "^1.24.0"
pyyaml = "^6.0.1"
langchain-openai = "^0.1.4"
fastapi = "^0.111.0"
streamlit = "^1.35.0"
uvicorn = "^0.30.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
